zion/jabber/requirements.txt
attrs==21.4.0
Automat==20.2.0
certifi==2021.10.8
charset-normalizer==2.0.11
constantly==15.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jabber-bot==0.1.1
paho-mqtt==1.6.1
pyTelegramBotAPI==4.4.0
requests==2.27.1
six==1.16.0
Twisted==22.1.0
twisted-iocpsupport==1.0.2
typing_extensions==4.1.0
urllib3==1.26.8
xmpppy==0.7.0
zope.interface==5.4.0
